@@ -1,5 +1,5 @@
-const { setNuxtConfig } = require('@opendreamnet/nuxtjs-base')
-const pkg = require('./package.json')
+import { setNuxtConfig, getNuxtConfig } from '@opendreamnet/nuxtjs-base'
+import pkg from './package.json'
 
 process.env.npm_package_name = pkg.name
 process.env.npm_package_description = pkg.description
@@ -7,15 +7,6 @@ process.env.npm_package_displayName = pkg.displayName
 process.env.npm_package_version = pkg.version
 
 export default setNuxtConfig({
-  // Global CSS: https://go.nuxtjs.dev/config-css
-  css: [
-  ],
-
-  // Plugins to run before rendering page: https://go.nuxtjs.dev/config-plugins
-  plugins: [
-    'node_modules/@opendreamnet/nuxtjs-base/plugins/boot.ts'
-  ],
-
   // Modules for dev and build (recommended): https://go.nuxtjs.dev/config-modules
   buildModules: [
     // https://go.nuxtjs.dev/typescript
@@ -23,7 +14,11 @@ export default setNuxtConfig({
     // https://go.nuxtjs.dev/tailwindcss
     '@nuxtjs/tailwindcss',
     // https://github.com/nuxt-community/style-resources-module
-    '@nuxtjs/style-resources'
+    '@nuxtjs/style-resources',
+    // https://github.com/nuxt-community/google-fonts-module
+    '@nuxtjs/google-fonts',
+    // https://github.com/nuxt-community/svg-sprite-module
+    '@nuxtjs/svg-sprite'
   ],
 
   // Modules: https://go.nuxtjs.dev/config-modules
@@ -31,20 +26,24 @@ export default setNuxtConfig({
     // https://go.nuxtjs.dev/pwa
     '@nuxtjs/pwa',
     // https://go.nuxtjs.dev/content
-    '@nuxt/content',
-    // https://github.com/nuxt-community/gtm-module
-    // ! Use v2.3.2: https://github.com/nuxt-community/gtm-module/issues/118
-    '@nuxtjs/gtm'
+    '@nuxt/content'
   ],
 
-  // Content module configuration: https://go.nuxtjs.dev/config-content
-  content: {},
+  plugins: [
+    ...getNuxtConfig().plugins!,
+    '~/plugins/api.ts'
+  ],
+
+  googleFonts: {
+    download: process.env.NODE_ENV === 'production',
+    families: {
+      Inter: [300, 400, 600, 700]
+    }
+  },
 
   //
   publicRuntimeConfig: {
-    githubURL: 'https://github.com/opendreamnet',
-    twitterURL: 'https://twitter.com/opendreamnetdev',
-    patreonURL: 'https://www.patreon.com/dreamnet'
+    opencollective: 'opendreamnet'
   },
 
   // Build Configuration: https://go.nuxtjs.dev/config-build
