@@ -1,11 +1,11 @@
-<template functional>
+<template>
   <div class="text-center">
-    <h1 class="mb-2 text-3xl md:leading-snug md:text-4xl title">
+    <h1 class="mb-2 text-3xl md:leading-snug md:text-5xl title">
       <span class="text-primary">Decentralized</span> applications<br>to fight <span class="text-danger">censorship</span>
     </h1>
 
-    <h2 class="text-lg text-snow-dark">
-      ...and other projects as well!
+    <h2 class="text-xl text-origin-dark">
+      and other projects
     </h2>
   </div>
 </template>
